+++ packages/core/src/Modals/View/ConfirmationComponent.php
@@ -47,6 +47,7 @@ class ConfirmationComponent extends Component
         public ?string $id = null,
         public ?string $closeAction = null,
         public ?int $zIndex = null,
+        public ?string $openOn = null,
     ) {
         $this->submitLabel ??= Trans::get('wire-core::actions.confirm_submit');
         $this->cancelLabel ??= Trans::get('wire-core::actions.confirm_cancel');